+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 - added stash navigation on ctrl + mousewheel
 - added support for the different poe.ninja urls based on the league and name
+- added latest_whisper placeholder for commands
+- added images for the items at the market
+- fixed an issue at the market enabling sockets to be toggleable
 
 ## 1.0.8 (2020-07-09)